refactor: Reader는 support/reader 패키지로 분리
