+++ src/main/java/umc/cockple/demo/domain/exercise/service/query/ExerciseRecommendationQueryService.java
@@ -8,10 +8,10 @@ import umc.cockple.demo.domain.exercise.converter.ExerciseConverter;
 import umc.cockple.demo.domain.exercise.domain.Exercise;
 import umc.cockple.demo.domain.exercise.dto.ExerciseRecommendationCalendarDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseRecommendationDTO;
-import umc.cockple.demo.domain.exercise.service.support.ExerciseBookmarkReader;
+import umc.cockple.demo.domain.exercise.service.support.reader.ExerciseBookmarkReader;
 import umc.cockple.demo.domain.exercise.service.support.ExerciseDistanceCalculator;
-import umc.cockple.demo.domain.exercise.service.support.ExerciseParticipantReader;
-import umc.cockple.demo.domain.exercise.service.support.ExerciseReader;
+import umc.cockple.demo.domain.exercise.service.support.reader.ExerciseParticipantReader;
+import umc.cockple.demo.domain.exercise.service.support.reader.ExerciseReader;
 import umc.cockple.demo.domain.member.domain.Member;
 import umc.cockple.demo.domain.member.domain.MemberAddr;
 import umc.cockple.demo.domain.member.service.support.MemberLookupService;
