+++ src/main/java/umc/cockple/demo/domain/exercise/service/support/reader/GuestReader.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.domain.exercise.service.support;
+package umc.cockple.demo.domain.exercise.service.support.reader;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
